Add files via upload
The three main models used are Logistic Regression,Random Forest and K-Nearest Neighbors for this classification problem. Hyperparameter Tuning is done to ghet optimal results and voting classifier is also used to  combine the predictions of three models: Logistic Regression, Random Forest, and K-Nearest Neighbors (KNN). The classifier operates in hard voting mode, where each model casts a "vote" for the predicted class (placed or not placed), and the class with the most votes is chosen as the final prediction.